Restore tests
