+++ internal/config/config.go
@@ -147,7 +147,7 @@ func Reset() {
 	viper.SetDefault(string(EventAggregatorRetryMaxDelay), "30s")
 	viper.SetDefault(string(EventDispatcherBufferLength), 5)
 	viper.SetDefault(string(EventDispatcherBatchTimeout), 0)
-	viper.SetDefault(string(EventTransportsEnabled), []string{"websocket"})
+	viper.SetDefault(string(EventTransportsEnabled), []string{"websockets"})
 	viper.SetDefault(string(HttpAddress), "127.0.0.1")
 	viper.SetDefault(string(HttpPort), 5000)
 	viper.SetDefault(string(HttpReadTimeout), "15s")
