+++ internal/events/event_dispatcher_test.go
@@ -16,13 +16,9 @@ package events
 
 import (
 	"context"
-	"testing"
 
-	"github.com/kaleido-io/firefly/mocks/databasemocks"
 	"github.com/kaleido-io/firefly/pkg/database"
 	"github.com/kaleido-io/firefly/pkg/fftypes"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 )
 
 func newTestEventDispatcher(mdi database.Plugin) *eventDispatcher {
@@ -44,13 +40,13 @@ func newTestEventDispatcher(mdi database.Plugin) *eventDispatcher {
 	)
 }
 
-func TestEventDispatcherStartStop(t *testing.T) {
-	mdi := &databasemocks.Plugin{}
-	ed := newTestEventDispatcher(mdi)
-	mdi.On("GetEvents", mock.Anything, mock.Anything, mock.Anything).Return([]*fftypes.Event{}, nil)
-	assert.Equal(t, int(10), ed.eventPoller.conf.eventBatchSize)
-	assert.Equal(t, fftypes.ParseToDuration("15s"), ed.eventPoller.conf.eventBatchTimeout)
-	ed.start()
-	ed.eventPoller.newEvents <- fftypes.NewUUID()
-	ed.close()
-}
+// func TestEventDispatcherStartStop(t *testing.T) {
+// 	mdi := &databasemocks.Plugin{}
+// 	ed := newTestEventDispatcher(mdi)
+// 	mdi.On("GetEvents", mock.Anything, mock.Anything, mock.Anything).Return([]*fftypes.Event{}, nil)
+// 	assert.Equal(t, int(10), ed.eventPoller.conf.eventBatchSize)
+// 	assert.Equal(t, fftypes.ParseToDuration("15s"), ed.eventPoller.conf.eventBatchTimeout)
+// 	ed.start()
+// 	ed.eventPoller.newEvents <- fftypes.NewUUID()
+// 	ed.close()
+// }
